requirements.txt
# RSS Feed Collection
feedparser==6.0.11
httpx==0.27.0

# NLP Processing
spacy==3.7.4
textblob==0.18.0.post0

# Machine Learning
scikit-learn==1.4.1.post1
numpy==1.26.4

# Google Cloud Services
google-cloud-firestore==2.14.0

# API Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
hypothesis==6.99.8
pytest-cov==4.1.0

# Development Tools
python-dotenv==1.0.1
